training_ner/requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
torchcrf>=1.1.0
tqdm>=4.65.0
pyyaml>=6.0
numpy>=1.24.0
scikit-learn>=1.3.0
